LayoutTests/editing/selection/user-select-all-parsing.html

<!DOCTYPE html>
    <head>
        <style>
        .userSelectAll {-webkit-user-select: all;}
        </style>
        <script src="../../fast/js/resources/js-test-pre.js"></script>
    </head>
    <body contentEditable="true">
        <div class="userSelectAll" id ="allArea">
            this is text of -webkit-user-select all
        </div>
        <div id="console"></div>
            
        <script>
            description("Test for Parsing new value all for -webkit-user-select");
            var userSelect = document.defaultView.getComputedStyle(document.getElementById("allArea"),null);
            shouldBeEqualToString('userSelect.getPropertyValue("-webkit-user-select")', 'all');
        </script>
        <script src="../../fast/js/resources/js-test-post.js"></script>
    </body>
</html>
