third_party/blink/web_tests/platform/generic/editing/selection/user-select-all-parsing-expected.txt

this is text of -webkit-user-select all

Test for Parsing new value all for -webkit-user-select

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS userSelect.getPropertyValue("-webkit-user-select") is "all"
PASS successfullyParsed is true

TEST COMPLETE
